WIP: Replace copy_out.sh with a bare container
Upstream ironic is proposing ironic-python-agent be able to consume
container-image packaged images directly. It is proposed that the
container image is an OCI image which only stores the qcow2 file instead
of a generic base image and extraction script also.

Once OSPRH-10995 is implemented, edpm-image-builder can be modified to
build a bare OCI container which only contains edpm-hardened-uefi.qcow2.

Jira: OSPRH-10996